requirements.txt
vectorbt
pandas
numpy
tqdm
tenacity
ratelimit
requests
python-dotenv
websockets>=10.0  # For WebSocket streaming 
duckdb>=0.9.0  # For high-performance analytics database
plotly  # For interactive charts
streamlit  # For dashboard interface
multiprocessing-logging  # For multiprocessing logging
pytest  # For testing
pytest-asyncio  # For async testing

# Real-time streaming and alerts
alpaca-trade-api>=3.0.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.20.0
redis>=4.5.0
celery>=5.2.0

# Email and notifications
aiosmtplib>=2.0.0
jinja2>=3.1.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Frontend build tools (for integrated deployment)
nodejs>=18.0.0 